Quote review: modified controller,service and repository

+++ src/main/java/it/proactivity/myinsurance/service/QuoteService.java
@@ -1,13 +1,15 @@
 package it.proactivity.myinsurance.service;
 
 import io.ebean.annotation.Transactional;
+import it.proactivity.myinsurance.exception.InvalidParamException;
+import it.proactivity.myinsurance.exception.InvalidQuoteException;
 import it.proactivity.myinsurance.exception.QuoteException;
 import it.proactivity.myinsurance.model.*;
+import it.proactivity.myinsurance.model.dto.*;
 import it.proactivity.myinsurance.repository.CarRepository;
 import it.proactivity.myinsurance.repository.HolderRepository;
 import it.proactivity.myinsurance.repository.OptionalExtraRepository;
 import it.proactivity.myinsurance.repository.QuoteRepository;
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -32,152 +34,156 @@ public class QuoteService extends MyInsuranceService{
     OptionalExtraRepository optionalExtraRepository;
 
     @Transactional
-    public List<Quote> findAll(){
-        try {
-            List<Quote> list = quoteRepository.findAll();
-            logger.debug("Quotes loaded correctly, tot: "+ list.size());
-            return list;
-        } catch (Exception e) {
-            logger.error("Error loading all quotes "+ e.getMessage());
-            return null;
+    public List<QuoteDTO> findAll() {
+
+        List<Quote> quotes = quoteRepository.findAll();
+        logger.debug("Quotes loaded correctly, tot: " + quotes.size());
+        List<QuoteDTO> quotesDTO = new ArrayList<>();
+
+        if (quotes == null)
+            return quotesDTO;
+        else {
+            quotesDTO = quotes.stream().sorted(Comparator.comparing(Quote::getDate))
+                    .map(quote -> {
+                        QuoteDTO quoteDTO = new QuoteDTO();
+                        quoteDTO.map(quote);
+                        return quoteDTO;
+                    }).collect(Collectors.toList());
         }
+
+        logger.debug(("quotes Found " + quotes.size()));
+        return quotesDTO;
     }
 
 
-    public Quote findById(Long id) {
-        try {
+    public QuoteDTO findById(Long id) {
+
             Quote quote = quoteRepository.findByIdWithHolderData(id);
-            if(quote != null)
-                 logger.debug("Quote Found: "+ quote.toString());
-            else
-                logger.debug("Quote is null ");
-            return quote;
-        } catch (Exception e) {
-            logger.error("Error finding the quote with id =" + id +"\n" + e.getMessage());
-            return null;
-        }
+
+            if(quote == null) {
+                logger.debug("No quote Found with the id = " + id);
+                return null;
+            }
+            else {
+                logger.debug("Quote Found: " + quote.toString());
+                QuoteDTO quoteDTO = new QuoteDTO();
+                quoteDTO.map(quote);
+                return quoteDTO;
+            }
+
     }
+
+
 
     public Quote findByQuoteNumber(String quoteNumber) {
-        try {
+
             Quote quote = quoteRepository.findByQuoteNumber(quoteNumber);
+
             if(quote != null)
                 logger.debug("Quote Found: "+ quote.toString());
             else
                 logger.debug("Quote is null ");
             return quote;
-        } catch (Exception e) {
-            logger.error("Error finding the quote with quoteNumber =" + quoteNumber +"\n" + e.getMessage());
-            return null;
-        }
+
     }
 
 
     public List<Quote> findByRegistrationMark(String registrationMark) {
-        try {
+
             List<Quote> list = quoteRepository.findByRegistrationMark(registrationMark);
+
             logger.debug("Quotes loaded correctly by registration mark, tot: "+ list.size());
             return list;
-        } catch (Exception e) {
-            logger.error("Error loading  quotes by registration mark"+ e.getMessage());
-            return null;
-        }
     }
 
 
     public List<Quote> findByHolderId(Long id) {
-        try {
+
             List<Quote> list = quoteRepository.findByHolderId(id);
             logger.debug("Quotes loaded correctly by holder id, tot: "+ list.size());
             return list;
-        } catch (Exception e) {
-            logger.error("Error loading  quotes by holder id"+ e.getMessage());
-            return null;
-        }
+
     }
+
 
     public List<QuoteByHolderAndCarDTO> findByHolderIdAndCar(QuoteByHolderAndCarDTO quoteByCarDTO) {
 
-        try {
-            List<QuoteByHolderAndCarDTO> quoteByCarDTOList = new ArrayList<>();
-            // get the list of the car
-            List<String> list = holderRepository.getRegistrationMarks(quoteByCarDTO.getHolder_id());
-            logger.debug("The Holder with this id " + quoteByCarDTO.getHolder_id() + " has these cars " + list);
-
-            list.forEach(reg -> {
-                QuoteByHolderAndCarDTO qByCarDTO = new QuoteByHolderAndCarDTO();
-                                 qByCarDTO.setHolder_id(quoteByCarDTO.getHolder_id());
-                                 qByCarDTO.setRegistrationMark(reg);
-                                 List<Quote> quotelist = quoteRepository.findByRegistrationMark(reg);
-                                 qByCarDTO.setQuotes(quotelist.stream().map(q -> {
-                                                                                   QuoteDTO qDTO = new QuoteDTO();
-                                                                                   qDTO.map(q);
-                                                                                   return qDTO;
-                                                                                 }).collect(Collectors.toList()));
-                                 quoteByCarDTOList.add(qByCarDTO);
-                                });
-
-            return quoteByCarDTOList;
-        } catch (Exception e) {
-            logger.error("Error loading  quotes by holder id"+ e.getMessage());
+        if(quoteByCarDTO == null) {
+            logger.error("QuoteByHolderAndCarDTO is null");
             return null;
         }
+
+        List<QuoteByHolderAndCarDTO> quoteByCarDTOList = new ArrayList<>();
+        // get the list of the car
+        List<String> list = holderRepository.getRegistrationMarks(quoteByCarDTO.getHolder_id());
+        logger.debug("The Holder with this id " + quoteByCarDTO.getHolder_id() + " has these cars " + list);
+
+        list.forEach(reg -> {
+            QuoteByHolderAndCarDTO qByCarDTO = new QuoteByHolderAndCarDTO();
+            qByCarDTO.setHolder_id(quoteByCarDTO.getHolder_id());
+            qByCarDTO.setRegistrationMark(reg);
+            List<Quote> quotelist = quoteRepository.findByRegistrationMark(reg);
+            qByCarDTO.setQuotes(quotelist.stream().map(q -> {
+                QuoteDTO qDTO = new QuoteDTO();
+                qDTO.map(q);
+                return qDTO;
+            }).collect(Collectors.toList()));
+            quoteByCarDTOList.add(qByCarDTO);
+        });
+
+        return quoteByCarDTOList;
+
     }
 
 
 
     public List<Quote> findByHolderIdAndRegistrationMark(Long holderId, String registrationMark) {
-        try {
             List<Quote> list = quoteRepository.findByHolderIdAndRegistrationMark(holderId, registrationMark);
             logger.debug("Quotes loaded correctly by holder id and registration mark, tot: "+ list.size());
             return list;
-        } catch (Exception e) {
-            logger.error("Error loading  quotes by holder id and registration mark"+ e.getMessage());
-            return null;
-        }
+
     }
 
 
-    public QuoteDTO save(QuoteForCreateDTO quoteForCreateDTO) throws QuoteException {
+    public QuoteDTO save(QuoteForCreateDTO quoteForCreateDTO)  {
+
+        if(quoteForCreateDTO == null){
+            logger.error("Input values are null");
+            throw new InvalidQuoteException("Input values are null");
+        }
 
         Holder holder;
         if((holder = holderRepository.findById(quoteForCreateDTO.getHolderId())) == null){
             logger.error("Holder is not registered. In order to receive our quotes, please register to the site");
-            throw new QuoteException("Holder is not registered. In order to receive our quotes, please register to the site");
+            throw new InvalidQuoteException("Quote is not registered. In order to receive our quotes, please register to the site");
         }
-
 
         if(!verifyRegistrationMarkPattern(quoteForCreateDTO.getRegistrationMark())){
             logger.error("Registration Mark is invalid");
-            throw new QuoteException("Registration Mark is invalid");
+            throw new InvalidQuoteException("Registration Mark is invalid");
         }
-
 
         if(!verifyRegistrationMarkNumber(quoteForCreateDTO.getRegistrationMark())){
             logger.error("You cannot create more than two quotes for the same registration mark ");
-            throw new QuoteException("You cannot create more than two quotes for the same registration mark ");
+            throw new InvalidQuoteException("You cannot create more than two quotes for the same registration mark ");
         }
-
 
         if(!verifyRegistrationDateCar(quoteForCreateDTO.getRegistrationDateCar())){
             logger.error("Registration date of the car is incorret ");
-            throw new QuoteException("Registration date of the car is incorret ");
+            throw new InvalidQuoteException("Registration date of the car is incorret ");
+        }
+        //create a new car or get the one from DB
+        Car car = getTheCar(holder,quoteForCreateDTO.getRegistrationMark(),
+                quoteForCreateDTO.getRegistrationDateCar(),
+                quoteForCreateDTO.getWorth());
+
+        if(car == null){
+            logger.error(" the car belongs to another holder " + quoteForCreateDTO.toString());
+            throw new InvalidQuoteException("the car belongs to another holder ");
         }
 
-        Quote quote = new Quote();
-        BeanUtils.copyProperties(quoteForCreateDTO,quote);
-
-        quote.setHolder(holder);
-        //add the car to the holder
-        Car car = new Car(holder,quoteForCreateDTO.getRegistrationMark(),quoteForCreateDTO.getRegistrationDateCar(),quoteForCreateDTO.getWorth());
-        car = carRepository.save(car);
-        //add the car to the quote
-        quote.setCar(car);
-        quote.setCost(calculateQuoteCost(quote.getCar().getWorth(),quote.getCar().getRegistrationDate(),quote.getPolicyType()));
-        quote.setDate(new Date(System.currentTimeMillis()));
-        quote.setQuoteNumber(quote.getHolder().getId() + "-"
-               + quote.getCar().getRegistrationMark() + "-"
-               + quoteRepository.getNextValidIndexByHolder(quote.getHolder().getId()));
+        //create a new quote with holder and car
+        Quote quote = getTheQuote(holder, car, quoteForCreateDTO.getPolicyType());
 
         quoteRepository.save(quote);
         QuoteDTO quoteDTO = new QuoteDTO();
@@ -187,15 +193,13 @@ public class QuoteService extends MyInsuranceService{
         return quoteDTO;
     }
 
-
-
-    public QuoteWithOptionalExtraDTO update(QuoteForUpdateDTO quoteForUpdateDTO) throws QuoteException {
+    public QuoteWithOptionalExtraDTO update(QuoteForUpdateDTO quoteForUpdateDTO)  {
 
         Quote quote = quoteRepository.findById(quoteForUpdateDTO.getId());
 
         if(quote == null) {
             logger.debug("cannot find Quote with the id: " + quoteForUpdateDTO.getId());
-            throw new QuoteException("cannot find Quote with the id: " + quoteForUpdateDTO.getId());
+            throw new InvalidQuoteException("cannot find Quote with the id: " + quoteForUpdateDTO.getId());
         }
 
         //remove previous extra
@@ -207,7 +211,7 @@ public class QuoteService extends MyInsuranceService{
 
             if(optionalExtra == null) {
                 logger.debug("Cannot find Optional extra with code " + code);
-                throw new QuoteException("cannot find Optional extra with code " + code);
+                throw new InvalidQuoteException("cannot find Optional extra with code " + code);
             }
             //A KASKO includes all the Optional Extras for this reason I clean up the
             //list of optional Extra and  leave only the KASKO
@@ -217,17 +221,16 @@ public class QuoteService extends MyInsuranceService{
                 break;
             }
 
-            quote.addOptionalExtra(optionalExtraRepository.findByCode(code));
+            quote.addOptionalExtra(optionalExtra);
             logger.debug("Added correctly new  optional extra : " + code );
         }
 
-        //recalculate the base costs
-        quote.setCost(calculateQuoteCost(quote.getCar().getWorth(),quote.getCar().getRegistrationDate(),quote.getPolicyType()));
-        //add the optional extras
-        if(quote.hasKasko())
-            quote.setCost(quote.getCost().add(addKaskoCosts(quote.getCar().getWorth(),quote.getCar().getRegistrationDate())).setScale(2, RoundingMode.CEILING));
-        else
-            quote.setCost(quote.getCost().add(addOptionalExtraCosts(quote.getCar().getWorth(),quote.getCar().getRegistrationDate(),quote.getOptionalExtras().size())).setScale(2, RoundingMode.CEILING));
+        //recalculate the cost
+        quote.setCost(calculateCost(quote.getCar().getWorth(),
+                quote.getCar().getRegistrationDate(),
+                quote.getPolicyType(),
+                quote.getOptionalExtras().size(),
+                quote.hasKasko()));
 
         quoteRepository.update(quote);
 
@@ -240,18 +243,30 @@ public class QuoteService extends MyInsuranceService{
 
 
 
-    public Long delete(QuoteForDeleteDTO quoteForDeleteDTO) throws QuoteException {
+    public Long delete(QuoteForDeleteDTO quoteForDeleteDTO)  {
+
+        if(quoteForDeleteDTO == null){
+            logger.error("input param is null");
+            throw new InvalidQuoteException("input param is null");
+        }
+
+        if(!quoteForDeleteDTO.isCorrect()){
+            logger.error("input param is null or incorrect" + quoteForDeleteDTO);
+            throw new InvalidQuoteException("input param is null or incorrect");
+        }
+
 
         Quote quote;
+
         if((quote = quoteRepository.findById(quoteForDeleteDTO.getId())) == null){
             logger.error("Quote is not registered.");
-            throw new QuoteException("Quote is not registered");
+            throw new InvalidQuoteException("Quote is not registered");
         }
 
 
         if(!verifyIfQuoteHasRegistrationMark(quoteForDeleteDTO.getId(),quoteForDeleteDTO.getRegistrationMark())){
-            logger.error("Quote has not this Registration Mark ");
-            throw new QuoteException("Quote has not this Registration Mark");
+            logger.error("Quote has not this Registration Mark or the registration Mark is assigned to another quote" + quoteForDeleteDTO);
+            throw new InvalidQuoteException("Quote has not this Registration Mark or the registration Mark is assigned to another quote");
         }
 
         quoteRepository.delete(quote);
@@ -260,8 +275,82 @@ public class QuoteService extends MyInsuranceService{
     }
 
 
+    private BigDecimal calculateCost(BigDecimal worth, Date registrationDate, PolicyType policyType, Integer totOptionalExtra, Boolean isKasko){
+        //recalculate the base costs
+         BigDecimal cost = calculateQuoteCost(worth,registrationDate,policyType);
+
+        //add the optional extras
+        if(isKasko)
+            cost = cost.add(addKaskoCosts(worth, registrationDate).setScale(2, RoundingMode.CEILING));
+        else
+            cost = cost.add(addOptionalExtraCosts(worth,registrationDate,totOptionalExtra).setScale(2, RoundingMode.CEILING));
+
+        return cost;
+    }
+
+    private Quote getTheQuote(Holder holder, Car car, PolicyType policyType){
+        Quote quote = new Quote();
+        quote.setPolicyType(policyType);
+        quote.setHolder(holder);
+        quote.setCar(car);
+        quote.setCost(calculateQuoteCost(quote.getCar().getWorth(),quote.getCar().getRegistrationDate(),quote.getPolicyType()));
+        quote.setDate(new Date(System.currentTimeMillis()));
+        quote.setQuoteNumber(quote.getHolder().getId() + "-"
+                + quote.getCar().getRegistrationMark() + "-"
+                + quoteRepository.getNextValidIndexByHolder(quote.getHolder().getId()));
+
+        return quote;
+    }
 
 
+    /**
+     *  if the car doesn't exist and doesn't own by another holder
+     *  I create a new car from the input data and
+     *  assign  the car to the holder
+     *  if the car exists, but doesn't belong to the holder
+     *  I return null
+     * @param holder
+     * @param registrationMark
+     * @param registrationDate
+     * @param worth
+     * @return
+     */
+    private Car getTheCar(Holder holder, String registrationMark,Date registrationDate, BigDecimal worth){
+
+        Car car = new Car(holder,registrationMark,
+                registrationDate,
+                worth);
+        //if the car exists, I check if belong to the holder
+        //if not, I return null
+        if (carRepository.existCar(registrationMark)){
+            if(holderHasThisCar(holder, registrationMark))
+                car = carRepository.findByRegistrationMark(registrationMark);
+            else
+                return null;
+        }
+        //If the car doesn't exist I create a new one
+        else
+            car = carRepository.save(car);
+
+        return car;
+    }
+
+    private  Boolean holderHasThisCar( Holder holder, String registrationMark)  {
+
+        if (carRepository.existCar(registrationMark)) {
+            if(getHolderIdOwnTheCar(registrationMark) == holder.getId()){
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private  Long getHolderIdOwnTheCar(String registrationMark)  {
+
+        if (carRepository.existCar(registrationMark))
+            return carRepository.findByRegistrationMark(registrationMark).getHolder().getId();
+        else return null;
+    }
 
     /**
      * check if the plate(registrationMark) has the right pattern match and
@@ -284,7 +373,7 @@ public class QuoteService extends MyInsuranceService{
             return true;
     }
 
-    private  Boolean verifyRegistrationDateCar(Date registrationDate) throws QuoteException {
+    private  Boolean verifyRegistrationDateCar(Date registrationDate)  {
 
         if(registrationDate.after(new Date(System.currentTimeMillis())))
             return false;
